requirements.txt
pyqt-darktheme==1.3.1
PyQt6==6.10.0
PyQt6-Qt6==6.10.0
PyQt6_sip==13.10.2
pyserial==3.5
